requirements.txt
Flask==1.1.2
Pillow==9.1.1
gunicorn>=20.1.0
gevent>=1.4